do an chuyen nganh cua tui

+++ admin/delete-brand.php
@@ -6,22 +6,38 @@
 
     $brand_id = $_GET['id'];
 
-    $sql = "delete from brands where brand_id = $brand_id";
+    $sql1 = "select *from products where product_brand = $brand_id";
 
-    $res = mysqli_query($con, $sql);
+    $res1 = mysqli_query($con, $sql1);
 
-    if($res == true) {
+    if($res1 == true)
+    {
+        $count = mysqli_num_rows($res1);
 
-        $_SESSION['delete'] = "Brand được xóa thành công!";
+        if($count > 0){
+            $_SESSION['delete'] = "Có khóa ngoại bên bảng product!";
 
-        header("location: brandlist.php");
+            header("location: brandlist.php");
 
-    } else {
+        }else {
+            $sql = "delete from brands where brand_id = $brand_id";
 
-        $_SESSION['delete'] = "Xóa không thành công. Vui lòng thử lại!";
+            $res = mysqli_query($con, $sql);
 
-        header("location: brandlist.php");
+            if($res == true) {
+
+                $_SESSION['delete'] = "Brand được xóa thành công!";
+
+                header("location: brandlist.php");
+
+             } else {
+
+                $_SESSION['delete'] = "Xóa không thành công. Vui lòng thử lại!";
+
+                header("location: brandlist.php");
 
         
+            }
+        }
     }
 ?>
